More progress and refactor
