+++ fss_check_config.yml
@@ -6,13 +6,14 @@ exclude_intervals:
 - 2022:293:16:00:00.000 2022:296:16:00:00.000
 - 2023:044:17:41:38.757 2023:048:03:49:31.349
 
+spm_pitch_limit: 135.0
+spm_pitch_warning: 137.0
+
 get_large_pitch_roll_error_intervals:
-  max_pitch: 136.0
-  max_err: 2.0
-  dt_join: 100
+  err_min: 2.0  # Minimum pitch / roll error
+  dt_join: 10  # Join adjacent intervals within this time (secs)
 
 plot_roll_pitch_vs_time:
-  pitch_max: 136
   plot_errs: false
 
-max_pitch_roll_time_plots: 4
+max_pitch_roll_time_plots: 10
